+++ app/vite.config.js
@@ -16,6 +16,7 @@ export default defineConfig({
         target: 'ws://localhost:3000',
         ws: true,
       },
+      '/outputs': 'http://localhost:3000',
     },
   },
 });